api/requirement.txt
amqp==2.5.2
appdirs==1.4.3
attrs==19.3.0
billiard==3.6.3.0
black==19.10b0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Django==2.2
django-celery-beat==2.0.0
django-timezone-field==4.0
djangorestframework==3.11.0
flower==0.9.4
humanize==0.5.1
idna==2.9
importlib-metadata==1.6.0
kombu==4.6.8
mysqlclient==1.4.6
pathspec==0.8.0
python-crontab==2.4.2
python-dateutil==2.8.1
pytz==2019.3
regex==2020.4.4
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
tornado==6.0.4
typed-ast==1.4.1
urllib3==1.25.9
vine==1.3.0
zipp==3.1.0
